pyproject.toml
[tool.poetry]
name = "typegql"
version = "4.0.2"
description = "A Python GraphQL library that makes use of type hinting and concurrency support with the new async/await syntax."
authors = ["Ciprian Tarta <ciprian@cipriantarta.ro>"]
license = "MIT"

readme = 'README.rst'  # Markdown files are supported

repository = "https://github.com/cipriantarta/typegql"
homepage = "https://github.com/cipriantarta/typegql"

keywords = ['graphql', 'python', 'asyncio', 'generics', 'typing']

classifiers=[
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development",
]

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = {version = "^3.4", optional = true}
graphql-core = ">=3"

[tool.poetry.dev-dependencies]
pytest-flake8 = "^1.0"
pytest-mypy = "^0.4.2"
sanic = "^19.9.0"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.8.1"
flake8-quotes = "^2.1.1"
bandit = "^1.6.2"
ipython = "^7.13.0"
pudb = "^2019.2"
black = "^19.10b0"
isort = "^4.3.21"

[tool.poetry.extras]
client = ["aiohttp", "cchardet"]

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
